Fixed A2

+++ anwarAssignment02/src/anwarAssignment02/AnwarAssignment02.java
@@ -16,8 +16,8 @@ public class AnwarAssignment02 {
      * @param args the command line arguments
      */
     public static void main(String[] args) {
-       int principal;
-       int rateofInterest;
+       double principal;
+       double rateofInterest;
        int yearTime;
        Scanner input = new Scanner(System.in);
        System.out.println("Welcome to the Simple Interest Calculator!");
@@ -25,12 +25,12 @@ public class AnwarAssignment02 {
                + "prompt and then click `Enter`");
        System.out.println("");
        System.out.println("Please enter in the Principal Amount:");
-       principal = input.nextInt();
+       principal = input.nextDouble();
        System.out.println("Please enter in the Rate of Interest:");     
-       rateofInterest = input.nextInt();
+       rateofInterest = input.nextDouble();
        System.out.println("Please enter in the Time (in years):");     
        yearTime = input.nextInt();
-       int simpleInterest = (principal*rateofInterest*yearTime)/100;
+       double simpleInterest = (principal*rateofInterest*yearTime)/100;
        System.out.println("The simple interest is: "+ simpleInterest);
     }
